Renamed `object` module to `document` in `lib`
