+++ lib/api/methods/fetchDPObjectsMethodFactory.js
@@ -1,10 +1,10 @@
 const InvalidParamsError = require('../InvalidParamsError');
-const InvalidWhereError = require('../../stateView/object/errors/InvalidWhereError');
-const InvalidOrderByError = require('../../stateView/object/errors/InvalidOrderByError');
-const InvalidLimitError = require('../../stateView/object/errors/InvalidLimitError');
-const InvalidStartAtError = require('../../stateView/object/errors/InvalidStartAtError');
-const InvalidStartAfterError = require('../../stateView/object/errors/InvalidStartAfterError');
-const AmbiguousStartError = require('../../stateView/object/errors/AmbiguousStartError');
+const InvalidWhereError = require('../../stateView/document/errors/InvalidWhereError');
+const InvalidOrderByError = require('../../stateView/document/errors/InvalidOrderByError');
+const InvalidLimitError = require('../../stateView/document/errors/InvalidLimitError');
+const InvalidStartAtError = require('../../stateView/document/errors/InvalidStartAtError');
+const InvalidStartAfterError = require('../../stateView/document/errors/InvalidStartAfterError');
+const AmbiguousStartError = require('../../stateView/document/errors/AmbiguousStartError');
 
 /**
  * @param {fetchDPObjects} fetchDPObjects
