+++ lib/app/ApiApp.js
@@ -25,9 +25,9 @@ const createCIDFromHash = require('../storage/stPacket/createCIDFromHash');
 const fetchContractFactory = require('../stateView/contract/fetchContractFactory');
 const fetchContractMethodFactory = require('../api/methods/fetchContractMethodFactory');
 
-const SVObjectMongoDbRepository = require('../stateView/object/SVObjectMongoDbRepository');
-const createSVObjectMongoDbRepositoryFactory = require('../stateView/object/createSVObjectMongoDbRepositoryFactory');
-const fetchDPObjectsFactory = require('../stateView/object/fetchDPObjectsFactory');
+const SVObjectMongoDbRepository = require('../stateView/document/SVObjectMongoDbRepository');
+const createSVObjectMongoDbRepositoryFactory = require('../stateView/document/createSVObjectMongoDbRepositoryFactory');
+const fetchDPObjectsFactory = require('../stateView/document/fetchDPObjectsFactory');
 const fetchDPObjectsMethodFactory = require('../api/methods/fetchDPObjectsMethodFactory');
 
 const getChainInfoFactory = require('../../lib/sync/info/chain/getChainInfoFactory');
