+++ lib/app/SyncApp.js
@@ -11,11 +11,11 @@ const sanitizer = require('../mongoDb/sanitizer');
 
 const SyncStateRepository = require('../sync/state/repository/SyncStateRepository');
 const SVContractMongoDbRepository = require('../stateView/contract/SVContractMongoDbRepository');
-const SVObjectMongoDbRepository = require('../stateView/object/SVObjectMongoDbRepository');
-const createSVObjectMongoDbRepositoryFactory = require('../stateView/object/createSVObjectMongoDbRepositoryFactory');
+const SVObjectMongoDbRepository = require('../stateView/document/SVObjectMongoDbRepository');
+const createSVObjectMongoDbRepositoryFactory = require('../stateView/document/createSVObjectMongoDbRepositoryFactory');
 const updateSVContractFactory = require('../stateView/contract/updateSVContractFactory');
-const updateSVObjectFactory = require('../stateView/object/updateSVObjectFactory');
-const revertSVObjectsForStateTransitionFactory = require('../stateView/object/revertSVObjectsForStateTransitionFactory');
+const updateSVObjectFactory = require('../stateView/document/updateSVObjectFactory');
+const revertSVObjectsForStateTransitionFactory = require('../stateView/document/revertSVObjectsForStateTransitionFactory');
 const revertSVContractsForStateTransitionFactory = require('../stateView/contract/revertSVContractsForStateTransitionFactory');
 const applyStateTransitionFactory = require('../stateView/applyStateTransitionFactory');
 const applyStateTransitionFromReferenceFactory = require('../stateView/applyStateTransitionFromReferenceFactory');
@@ -27,7 +27,7 @@ const createStateTransitionsFromBlockFactory = require('../blockchain/createStat
 const readBlockchainFactory = require('../blockchain/reader/readBlockchainFactory');
 
 const fetchContractFactory = require('../stateView/contract/fetchContractFactory');
-const fetchDPObjectsFactory = require('../stateView/object/fetchDPObjectsFactory');
+const fetchDPObjectsFactory = require('../stateView/document/fetchDPObjectsFactory');
 
 const STPacketIpfsRepository = require('../storage/stPacket/STPacketIpfsRepository');
