+++ lib/stateView/document/SVObjectMongoDbRepository.js
@@ -2,7 +2,7 @@ const bs58 = require('bs58');
 
 const DPObject = require('@dashevo/dpp/lib/document/Document');
 
-const SVObject = require('../../stateView/object/SVObject');
+const SVObject = require('./SVObject');
 const Reference = require('../revisions/Reference');
 
 const createRevisions = require('../revisions/createRevisions');
