+++ lib/stateView/document/createSVObjectMongoDbRepositoryFactory.js
