+++ lib/stateView/document/errors/AmbiguousStartError.js
