+++ lib/stateView/document/errors/InvalidLimitError.js
