+++ lib/stateView/document/errors/InvalidStartAfterError.js
