+++ lib/stateView/document/errors/InvalidWhereError.js
