+++ lib/stateView/document/fetchDPObjectsFactory.js
