+++ lib/stateView/document/revertSVObjectsForStateTransitionFactory.js
