+++ lib/stateView/document/updateSVObjectFactory.js
