+++ lib/sync/attachSyncLogger.js
@@ -1,6 +1,6 @@
 const ReaderMediator = require('../blockchain/reader/BlockchainReaderMediator');
 
-const SVObject = require('../stateView/object/SVObject');
+const SVObject = require('../stateView/document/SVObject');
 
 const WrongSequenceError = require('../blockchain/reader/eventHandlers/errors/WrongSequenceError');
 const NotAbleToValidateSequenceError = require('../blockchain/reader/eventHandlers/errors/NotAbleToValidateSequenceError');
