+++ lib/test/fixtures/getSVObjectsFixture.js
@@ -1,4 +1,4 @@
-const SVObject = require('../../stateView/object/SVObject');
+const SVObject = require('../../stateView/document/SVObject');
 
 const getDPObjectsFixture = require('./getDPObjectsFixture');
 const getReferenceFixture = require('./getReferenceFixture');
